Changed printf so that time is a float.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@1886 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/io/StatsProbe.cpp
@@ -73,11 +73,11 @@ int StatsProbe::outputState(float time, PVLayer * l)
 
    if (type == BufActivity) {
       float freq = 1000.0 * (sum/nk);
-      fprintf(fp, "%st==%4d N==%d Total==%f Min==%f Avg==%f Hz (/dt ms) Max==%f\n", msg, (int)time,
+      fprintf(fp, "%st==%6.1f N==%d Total==%f Min==%f Avg==%f Hz (/dt ms) Max==%f\n", msg, time,
               nk, (float)sum, fMin, freq, fMax);
    }
    else {
-      fprintf(fp, "%st==%4d N==%d Total==%f Min==%f Avg==%f Max==%f\n", msg, (int)time,
+      fprintf(fp, "%st==%6.1f N==%d Total==%f Min==%f Avg==%f Max==%f\n", msg, time,
               nk, (float)sum, fMin, (float)(sum / nk), fMax);
    }
